Bug fixes (autocomplete and clearing code)
