+++ src/components/collab_elements/languageOptions.js
@@ -61,12 +61,6 @@ export const languageOptions = [
       value: "perl",
     },
     {
-      id: 70,
-      name: "Python (2.7.17)",
-      label: "Python (2.7.17)",
-      value: "python",
-    },
-    {
       id: 71,
       name: "Python (3.8.1)",
       label: "Python (3.8.1)",
